Do not use tree.lp for config editor
